refactor : 부스 패키지 이동
