+++ src/main/java/likelion/festival/booth/application/dto/BoothFilterDto.java
@@ -1,7 +1,7 @@
-package likelion.festival.dto;
+package likelion.festival.booth.application.dto;
 
 import com.sun.istack.NotNull;
-import likelion.festival.entity.BoothType;
+import likelion.festival.booth.domain.BoothType;
 import likelion.festival.entity.Image;
 import lombok.Builder;
 import lombok.Data;
