+++ src/main/java/likelion/festival/dto/ImageDto.java
@@ -1,7 +1,7 @@
 package likelion.festival.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import likelion.festival.entity.Booth;
+import likelion.festival.booth.domain.Booth;
 import likelion.festival.entity.Image;
 import likelion.festival.entity.Notification;
 import lombok.*;
