+++ src/main/java/likelion/festival/dto/MenuRequestDto.java
@@ -1,6 +1,6 @@
 package likelion.festival.dto;
 
-import likelion.festival.entity.Booth;
+import likelion.festival.booth.domain.Booth;
 import lombok.*;
 
 @Data
