+++ src/main/java/likelion/festival/entity/Likes.java
@@ -1,6 +1,7 @@
 package likelion.festival.entity;
 
 import com.sun.istack.NotNull;
+import likelion.festival.booth.domain.Booth;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
