+++ src/main/java/likelion/festival/service/CommentService.java
@@ -4,12 +4,12 @@ package likelion.festival.service;
 import likelion.festival.dto.CommentPasswordDto;
 import likelion.festival.dto.CommentRequestDto;
 import likelion.festival.dto.CommentResponseDto;
-import likelion.festival.entity.Booth;
+import likelion.festival.booth.domain.Booth;
 import likelion.festival.entity.Comment;
 import likelion.festival.exception.WrongBoothId;
 import likelion.festival.exception.WrongCommentId;
 import likelion.festival.exception.WrongPassword;
-import likelion.festival.repository.BoothRepository;
+import likelion.festival.booth.domain.repository.BoothRepository;
 import likelion.festival.repository.CommentRepository;
 import likelion.festival.security.Encrypt;
 import lombok.RequiredArgsConstructor;
